auth_service/requirements.txt
channels==4.1.0
daphne==4.1.2
channels-redis==4.0.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-cors-headers==4.6.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
phonenumbers==8.13.49
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.2
typing-extensions==4.12.2
urllib3==2.2.3
psycopg2-binary>=2.9.1